+++ app/src/main/java/com/paisalo/newinternalsourcingapp/Fragments/home/HomeFragment.java
@@ -75,7 +75,6 @@ public class HomeFragment extends Fragment {
         View root = binding.getRoot();
         binding.backgroundImageView.setVisibility(View.GONE);
 
-        Toast.makeText(getActivity(), "STarting", Toast.LENGTH_LONG).show();
         imageView = root.findViewById(R.id.imageTop);
         videoView = root.findViewById(R.id.videoViewBanner);
